Update get IPN response

+++ src/PlayTogether.Core/Dtos/Incoming/Momo/MomoIPNRequest.cs
@@ -5,14 +5,14 @@ namespace PlayTogether.Core.Dtos.Incoming.Momo
         public string PartnerCode { get; set; }
         public string OrderId { get; set; }
         public string RequestId { get; set; }
-        public float Amount { get; set; }
+        public long Amount { get; set; }
         public string OrderInfo { get; set; }
         public string OrderType { get; set; }
-        public double TransId { get; set; }
+        public long TransId { get; set; }
         public int ResultCode { get; set; }
         public string Message { get; set; }
         public string PayType { get; set; }
-        public double ResponseTime { get; set; }
+        public long ResponseTime { get; set; }
         public string ExtraData { get; set; }
         public string Signature { get; set; }
     }
